refact, tests: work on splitting tests into separate packages
