+++ taf/tests/test_updater/test_handlers/test_types.py
